refactor: Move gateway request for scaleup in scaling package
