+++ zeebe/broker/src/main/java/io/camunda/zeebe/broker/partitioning/scaling/BrokerClientPartitionScalingExecutor.java
@@ -12,9 +12,9 @@ import io.camunda.zeebe.broker.client.api.BrokerErrorException;
 import io.camunda.zeebe.broker.client.api.BrokerRejectionException;
 import io.camunda.zeebe.broker.client.api.dto.BrokerResponse;
 import io.camunda.zeebe.dynamic.config.changes.PartitionScalingChangeExecutor;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerPartitionBootstrappedRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.BrokerPartitionScaleUpRequest;
-import io.camunda.zeebe.gateway.impl.broker.request.GetScaleUpProgress;
+import io.camunda.zeebe.gateway.impl.broker.request.scaling.BrokerPartitionBootstrappedRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.scaling.BrokerPartitionScaleUpRequest;
+import io.camunda.zeebe.gateway.impl.broker.request.scaling.GetScaleUpProgress;
 import io.camunda.zeebe.protocol.impl.record.value.scaling.ScaleRecord;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.scheduler.ConcurrencyControl;
