+++ zeebe/gateway/src/main/java/io/camunda/zeebe/gateway/impl/broker/request/scaling/BrokerPartitionScaleUpRequest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.impl.broker.request;
+package io.camunda.zeebe.gateway.impl.broker.request.scaling;
 
 import io.camunda.zeebe.broker.client.api.dto.BrokerExecuteCommand;
 import io.camunda.zeebe.protocol.Protocol;
@@ -15,13 +15,13 @@ import io.camunda.zeebe.protocol.record.intent.scaling.ScaleIntent;
 import io.camunda.zeebe.util.buffer.BufferWriter;
 import org.agrona.DirectBuffer;
 
-public class BrokerPartitionBootstrappedRequest extends BrokerExecuteCommand<ScaleRecord> {
+public final class BrokerPartitionScaleUpRequest extends BrokerExecuteCommand<ScaleRecord> {
 
   private final ScaleRecord requestDto = new ScaleRecord();
 
-  public BrokerPartitionBootstrappedRequest(final int bootstrappedPartition) {
-    super(ValueType.SCALE, ScaleIntent.MARK_PARTITION_BOOTSTRAPPED);
-    requestDto.markPartitionBootstrapped(bootstrappedPartition);
+  public BrokerPartitionScaleUpRequest(final int desiredPartitionCount) {
+    super(ValueType.SCALE, ScaleIntent.SCALE_UP);
+    requestDto.scaleUp(desiredPartitionCount);
 
     // set the target partition for this request
     setPartitionId(Protocol.DEPLOYMENT_PARTITION);
